feat: add direct pin mod
